agrego confirmacion al eliminar genero

+++ controllers/genre_controller.php
@@ -24,31 +24,33 @@ class genre_controller{
     }
 
     public function update_genre($id){           //CONTROLADOR - MODIFICA/EDITA GÉNERO YA CARGADO
-        
         if (isset($_POST['name_genre'], $_POST['description_genre'])){
             $name_genre = $_POST['name_genre'];
             $description_genre = $_POST['description_genre'];
             $this->model->update_genre_model($name_genre, $description_genre, $id);
-            // $this->view->show_genres();
-            // header('Location: generos');
             header("Location: ". BASE_URL . "generos");                                        //UNA VEZ MODIFICADO EL REGISTRO, REDIRECCIONO A LA PAGINA DE GENEROS
         }
     }
    
-
     public function add_genre(){           //Funcion para agregar un genero nuevo. Lee los valores que ingreso el usuario en el formulario
         if (isset($_POST['name_genre'], $_POST['description_genre'])){
             $name_genre=$_REQUEST['name_genre'];
             $description_genre=$_REQUEST['description_genre'];
-                    
             $this->model->insert_genre($name_genre, $description_genre);
             header("Location: generos");                              //UNA VEZ QUE AGREGO EL GENERO NUEVO, REDIRECCIONO A LA PAGINA DE GENEROS
         } 
     }
 
-    public function delete_genre($id){   //Funcion para eliminar un juego segun su id
-        $this->model->delete_genre($id);
-        header("Location: ". BASE_URL . "generos");
+    public function delete_genre($id){   
+    // Comprobar si se envió la confirmación desde el cliente
+        if (isset($_POST['confirmacion']) && $_POST['confirmacion'] === 'confirmado') {
+            // Realizar la eliminación del género
+            $this->model->delete_genre($id);
+            header("Location: " . BASE_URL . "generos");
+        } else {
+            // Mostrar la página de confirmación al usuario
+            $this->view->mostrar_confirmacion($id);
+        }
     }
-    
+
 }
